+++ spec/integration/rails/support/home_controller.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 class HomeController < ActionController::Base
-  def index
+  def index1
     request_context = ::Castle::Context::Prepare.call(request)
-    payload = ::Castle::Client.to_options(
+    payload = {
       event: '$login.succeeded',
       user_id: '123',
       properties: {
@@ -12,9 +12,48 @@ class HomeController < ActionController::Base
       user_traits: {
         key: 'value'
       }
-    )
+    }
     client = ::Castle::Client.new(context: request_context)
     client.track(payload)
+
+    render inline: 'hello'
+  end
+
+  def index2
+    payload = ::Castle::Payload::Prepare.call(
+      {
+        event: '$login.succeeded',
+        user_id: '123',
+        properties: {
+          key: 'value'
+        },
+        user_traits: {
+          key: 'value'
+        }
+      },
+      request
+    )
+    client = ::Castle::Client.new
+    client.track(payload)
+
+    render inline: 'hello'
+  end
+
+  def index3
+    payload = ::Castle::Payload::Prepare.call(
+      {
+        event: '$login.succeeded',
+        user_id: '123',
+        properties: {
+          key: 'value'
+        },
+        user_traits: {
+          key: 'value'
+        }
+      },
+      request
+    )
+    Castle::API::Track.call(payload)
 
     render inline: 'hello'
   end