+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - [#213](https://github.com/castle/castle-ruby/pull/213) rename config.url to config.base_url
 - [#214](https://github.com/castle/castle-ruby/pull/214) reorganize structure of the SDK, Castle::API::Session renamed to Castle::Session
+- [#216](https://github.com/castle/castle-ruby/pull/216) added new context and payload builders, changed DSL
 
 ## 5.0.0 (2020-09-29)